hrtimer: Change cpu_relax() in hrtimer_cancel() to udelay(1) instead.
There is a lock contention in the case when we execute run_hrtimer and
the hrtimer's callback wakes up a thread which tries to cancel the
timer.  The thread trying to cancel the timer fails and retries because
the hrtimer handler is active but the thread also casues lock contention
which inhibits the hrtimer handler from making forward progress.

Bug: 79694843
Test: manual - flash device and check using a trace.

Change-Id: I366f0ec178fa23d38788ca5b93ded88056b6c552
Signed-off-by: Miguel de Dios <migueldedios@google.com>
Suggested by: Martin Liu <liumartin@google.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>