iommu: msm: Rewrite to improve clarity and performance
This scope of this driver's lock usage is extremely wide, leading to
excessively long lock hold times. Additionally, an entire linked list
is traversed for the sole purpose of trying to find a reason to invoke a
BUG. These are the two most significant contributors to poor performance
in this driver.

Fix all of this by greatly reducing the scope of the locks used and by
using atomic reader/writer locks. The superfluous linked list traversal
is also gone, and the entire driver's code is refactored and improved
for better cleanliness and clarity.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>